+++ src/i18n/index.ts
@@ -4,6 +4,7 @@ import { register } from "timeago.js"
 import LanguageDetector from 'i18next-browser-languagedetector';
 import set from "lodash/set";
 
+import store from "./../store";
 
 import enTranslations from './locales/en.json';
 import frTranslations from './locales/fr.json';
@@ -12,6 +13,7 @@ import deTranslations from './locales/de.json';
 import ruTranslations from './locales/ru.json';
 import ptbrTranslations from './locales/ptbr.json';
 import esTranslations from './locales/es.json';
+import uaTranslations from './locales/ua.json';
 
 
 import timePl from "timeago.js/lib/lang/pl";
@@ -20,6 +22,7 @@ import timeDe from "timeago.js/lib/lang/de";
 import timeRu from "timeago.js/lib/lang/ru";
 import timePtBr from "timeago.js/lib/lang/pt_BR";
 import timeEs from "timeago.js/lib/lang/es";
+import timeUa from "timeago.js/lib/lang/uk";
 
 
 register("pl", timePl);
@@ -28,6 +31,7 @@ register("de", timeDe);
 register("ru", timeRu);
 register("ptbr", timePtBr);
 register("es", timeEs);
+register("ua", timeUa);
 
 
 
@@ -38,27 +42,34 @@ export const resources = {
     de: deTranslations as ResourceLanguage,
     ru: ruTranslations as ResourceLanguage,
     ptbr: ptbrTranslations as ResourceLanguage,
-    es: esTranslations as ResourceLanguage
+    es: esTranslations as ResourceLanguage,
+    ua: uaTranslations as ResourceLanguage,
 } as const;
 
-declare let window:Record<string, unknown>;
+declare let window: Record<string, unknown>;
 window.missing = {};
 
+const blacklistedNamespaces = ['localeNames'];
 const missingKeyHandler = (lngs: string[], ns: string, key: string, fallbackValue: string) => {
-    // eslint-disable-next-line @typescript-eslint/ban-types
-    set(window.missing as object, [ns, key].join('.'), fallbackValue);
-    //then use `copy(window.missing)` in chrome dev tools console
+
+    if (!blacklistedNamespaces.includes(ns)) {
+
+        // eslint-disable-next-line @typescript-eslint/ban-types
+        set(window.missing as object, [ns, key].join('.'), fallbackValue);
+        store.setState({ missingTranslations: window.missing as Map<string, unknown> });
+        //then use `copy(window.missing)` in chrome dev tools console
+    }
 }
 const debug = process.env.NODE_ENV !== 'production'
 i18n
     .use(LanguageDetector)
     .use(initReactI18next)
     .init({
-        fallbackLng: 'en',
+        // fallbackLng: 'en',
         debug,
         resources,
         ns: Object.keys(enTranslations),
-        saveMissing: debug,
+        saveMissing: true,
         missingKeyHandler
     });
 export default i18n;